split('') will now split a string at every character.
Previously, this was an undefined case that would result in an infinite
loop. Now it acts the same as in JavaScript, and it returns an list
where every element is 1 character of the string.

This is useful for applying list methods to strings such as map, reduce,
filter, contains etc